Take care of review comments. Going back to finagle 6.41. finagle-http-compat has been removed in more recent versions, taking care of that may need non-trivial changes. Will upgrade finagle version in a follow up change.

+++ storehaus-redis/src/main/scala/com/twitter/storehaus/redis/RedisSortedSetStore.scala
@@ -39,7 +39,7 @@ class RedisSortedSetStore(client: Client)
   /** Returns the whole set as a tuple of seq of (member, score).
    *  An empty set is represented as None. */
   override def get(k: Buf): Future[Option[Seq[(Buf, Double)]]] =
-    client.zRange(k, 0.toLong, -1.toLong, true).map(
+    client.zRange(k, 0L, -1L, true).map(
       _.left.toOption.map( _.asTuples).filter(_.nonEmpty)
     )
